Fixed issue with WaitUntilRemoved when used like this
browser.Elements[0].WaitUntilRemoved.

Issue is caused by calling Exists which calls FindNativeElement, which clears _cachedNativeElement. For these collection elements there is no elementFinder so an exception gets thrown as soon as the element is removed. Need to reimplement caching in Element for the two different ways an element can be set/found (basically for the two different constructors there should be a different caching mechanism.)

git-svn-id: svn://svn.code.sf.net/p/watin/code/trunk@1093 6ca19b5b-dac0-4595-9ef0-4637b6ca95b2